tools/visual-pipeline-and-platform-evaluation-tool/requirements.txt
gradio==5.36.2
numpy==2.2.6
opencv-python-headless==4.12.0.88
openvino==2025.1.0
pandas==2.3.0
plotly==6.0.1
psutil==6.1.1
pyyaml==6.0.2
requests==2.32.4
fastapi[standard]==0.118.0